commmit

+++ Lecture2/src/if_else/FtoC.java
@@ -4,7 +4,7 @@ public class FtoC {
     public static void main(String args[]) {
         Scanner s = new Scanner(System.in);
         int f = s.nextInt();
-        int cel = (int)((5.0 / 9) * (f - 32));
+        float cel = (float)((5.0 / 9) * (f - 32));
         System.out.println(cel);
     }
 }
